+++ not_for_release/testFramework/Feature/TestTaxProductBasisStoreShippingTaxableBasisShipping/BasicTaxStoreWithShippingTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Feature\TestStoreTaxCalculationsWithShippingTax;
+namespace Feature\TestTaxProductBasisStoreShippingTaxableBasisShipping;
 
 use Tests\Support\zcFeatureTestCaseStore;
